refactor GetCategoryProgression
