+++ Src/MoneyFox.Core.Tests/Queries/Statistics/GetCategoryProgression/GetCategoryProgressionQueryTests.cs
@@ -1,0 +1,32 @@
+﻿namespace MoneyFox.Core.Tests.Queries.Statistics.GetCategoryProgression;
+
+using FluentAssertions;
+using MoneyFox.Core.Queries.Statistics;
+
+public class GetCategoryProgressionQueryTests
+{
+    [Fact]
+    public void ThrowsException_WhenStartDateIsAfterEndDate()
+    {
+        // Act
+        var act = () => new GetCategoryProgression.Query(3, startDate: DateOnly.FromDateTime(DateTime.Today).AddDays(1), endDate: DateOnly.FromDateTime(DateTime.Today));
+
+        // Assert
+        act.Should().Throw<InvalidDateRangeException>();
+    }
+
+    [Fact]
+    public void CreateQueryAndAssignValues()
+    {
+        // Arrange
+        var categoryId = 3;
+
+        // Act
+        var query = new GetCategoryProgression.Query(categoryId, startDate: DateOnly.FromDateTime(DateTime.Today), endDate: DateOnly.FromDateTime(DateTime.Today));
+
+        // Assert
+        query.CategoryId.Should().Be(categoryId);
+        query.StartDate.Should().Be(DateOnly.FromDateTime(DateTime.Today));
+        query.EndDate.Should().Be(DateOnly.FromDateTime(DateTime.Today));
+    }
+}
